SSCS-4106: Evidence received Appellant notification (paper case) (#353)
* SSCS-4106: Evidence received (not attending a hearing) | Appellant notification

* put in sms acceptance criteria
