+++ src/e2e/java/uk/gov/hmcts/reform/sscs/functional/sya/notifications/WithRepresentativePersonalisationTest.java
@@ -1,17 +1,13 @@
 package uk.gov.hmcts.reform.sscs.functional.sya.notifications;
 
 import static org.junit.Assert.assertTrue;
-import static uk.gov.hmcts.reform.sscs.config.AppealHearingType.ORAL;
-import static uk.gov.hmcts.reform.sscs.config.AppealHearingType.PAPER;
 import static uk.gov.hmcts.reform.sscs.domain.notify.NotificationEventType.*;
 
 import java.lang.reflect.Field;
 import java.util.List;
 import junitparams.Parameters;
-import org.apache.commons.lang3.StringUtils;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Value;
-import uk.gov.hmcts.reform.sscs.config.AppealHearingType;
 import uk.gov.hmcts.reform.sscs.domain.notify.NotificationEventType;
 import uk.gov.hmcts.reform.sscs.functional.AbstractFunctionalTest;
 import uk.gov.service.notify.Notification;
@@ -41,19 +37,6 @@ public class WithRepresentativePersonalisationTest extends AbstractFunctionalTes
     private String hearingBookedRepsEmailId;
     @Value("${notification.hearingBooked.representative.smsId}")
     private String hearingBookedRepsSmsId;
-
-    @Value("${notification.evidenceReceived.appellant.emailId}")
-    private String evidenceReceivedAppellantEmailId;
-    @Value("${notification.evidenceReceived.appellant.smsId}")
-    private String evidenceReceivedAppellantSmsId;
-    @Value("${notification.paper.evidenceReceived.representative.emailId}")
-    private String paperEvidenceReceivedRepsEmailId;
-    @Value("${notification.evidenceReceived.representative.smsId}")
-    private String paperEvidenceReceivedRepsSmsId;
-    @Value("${notification.oral.evidenceReceived.representative.emailId}")
-    private String oralEvidenceReceivedRepsEmailId;
-    @Value("${notification.evidenceReceived.representative.smsId}")
-    private String oralEvidenceReceivedRepsSmsId;
     @Value("${notification.appealCreated.appellant.emailId}")
     private String appealCreatedAppellantEmailId;
     @Value("${notification.appealCreated.appellant.smsId}")
@@ -123,34 +106,6 @@ public class WithRepresentativePersonalisationTest extends AbstractFunctionalTes
     }
 
     @Test
-    @Parameters(method = "evidenceReceivedNotifications")
-    public void givenEventAndRepsSubscription_shouldSendNotificationToReps(AppealHearingType appealHearingType,
-                                                                           NotificationEventType notificationEventType)
-            throws Exception {
-        //Given
-        final String appellantEmailId = getFieldValue(appealHearingType, notificationEventType, "AppellantEmailId");
-        final String appellantSmsId = getFieldValue(appealHearingType, notificationEventType, "AppellantSmsId");
-        final String repsEmailId = getFieldValue(appealHearingType, notificationEventType, "RepsEmailId");
-        final String repsSmsId = getFieldValue(appealHearingType, notificationEventType, "RepsSmsId");
-
-        simulateCcdCallback(notificationEventType,
-                "representative/"
-                        + (appealHearingType != null ? (appealHearingType.name().toLowerCase() + "-") : "")
-                        + notificationEventType.getId() + "Callback.json");
-
-        List<Notification> notifications = tryFetchNotificationsForTestCase(
-                appellantEmailId, appellantSmsId,
-                repsEmailId, repsSmsId);
-
-        assertNotificationBodyContains(notifications, appellantEmailId);
-        assertNotificationBodyContains(notifications, appellantSmsId);
-
-        String representativeName = "Harry Potter";
-        assertNotificationBodyContains(notifications, repsEmailId, representativeName);
-        assertNotificationBodyContains(notifications, repsSmsId);
-    }
-
-    @Test
     @Parameters(method = "eventTypeAndSubscriptions")
     public void givenEventAndNoRepsSubscription_shouldNotSendNotificationToReps(NotificationEventType notificationEventType)
         throws Exception {
@@ -172,39 +127,6 @@ public class WithRepresentativePersonalisationTest extends AbstractFunctionalTes
         List<Notification> notificationsNotFound = tryFetchNotificationsForTestCaseWithFlag(true,
             repsEmailId, repsSmsId);
         assertTrue(notificationsNotFound.isEmpty());
-    }
-
-    @Test
-    @Parameters(method = "evidenceReceivedNotifications")
-    public void givenEventAndNoRepsSubscription_shouldNotSendNotificationToReps(AppealHearingType appealHearingType,
-                                                                                NotificationEventType notificationEventType)
-            throws Exception {
-
-        final String appellantEmailId = getFieldValue(appealHearingType, notificationEventType, "AppellantEmailId");
-        final String appellantSmsId = getFieldValue(appealHearingType, notificationEventType, "AppellantSmsId");
-        final String repsEmailId = getFieldValue(appealHearingType, notificationEventType, "RepsEmailId");
-        final String repsSmsId = getFieldValue(appealHearingType, notificationEventType, "RepsSmsId");
-
-        simulateCcdCallback(notificationEventType,
-                "representative/" + "no-reps-subscribed-"
-                        + (appealHearingType != null ? (appealHearingType.name().toLowerCase() + "-") : "")
-                        + notificationEventType.getId() + "Callback.json");
-
-        List<Notification> notifications = tryFetchNotificationsForTestCase(appellantEmailId,
-                appellantSmsId);
-        assertNotificationBodyContains(notifications, appellantEmailId);
-        assertNotificationBodyContains(notifications, appellantSmsId);
-
-        List<Notification> notificationsNotFound = tryFetchNotificationsForTestCaseWithFlag(true,
-                repsEmailId, repsSmsId);
-        assertTrue(notificationsNotFound.isEmpty());
-    }
-
-    private Object[] evidenceReceivedNotifications() {
-        return new Object[]{
-            new Object[]{ORAL, EVIDENCE_RECEIVED_NOTIFICATION},
-            new Object[]{PAPER, EVIDENCE_RECEIVED_NOTIFICATION},
-        };
     }
 
     public void givenHearingPostponedEventAndRepsSubscription_shouldSendEmailOnlyNotificationToReps()
@@ -257,30 +179,7 @@ public class WithRepresentativePersonalisationTest extends AbstractFunctionalTes
         return fieldValue;
     }
 
-    private String getFieldValue(AppealHearingType appealHearingType, NotificationEventType notificationEventType, String fieldName) throws Exception {
-        Field field = null;
-
-        if (appealHearingType != null) {
-            field = getField(appealHearingType, notificationEventType, fieldName);
-        }
-
-        if (field == null) {
-            field = this.getClass().getDeclaredField(notificationEventType.getId() + fieldName);
-        }
-
-        field.setAccessible(true);
-        return (String) field.get(this);
-    }
-
-    private Field getField(AppealHearingType appealHearingType, NotificationEventType notificationEventType, String fieldName) throws NoSuchFieldException {
-        try {
-            return this.getClass().getDeclaredField(appealHearingType.name().toLowerCase() + StringUtils.capitalize(notificationEventType.getId()) + fieldName);
-        } catch (NoSuchFieldException e) {
-            return null;
-        }
-    }
-
-
+    @SuppressWarnings({"Indentation", "unused"})
     private Object[] eventTypeAndSubscriptions() {
         return new Object[]{
             new Object[]{APPEAL_LAPSED_NOTIFICATION},
