+++ src/main/java/uk/gov/hmcts/reform/sscs/config/NotificationConfig.java
@@ -2,6 +2,8 @@ package uk.gov.hmcts.reform.sscs.config;
 
 import java.util.Locale;
 import javax.validation.constraints.NotNull;
+
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.env.Environment;
@@ -84,7 +86,7 @@ public class NotificationConfig {
         if (templateId == null) {
             templateId = env.getProperty("notification." + templateName + "." + notificationType);
         }
-        return templateId;
+        return StringUtils.stripToNull(templateId);
     }
 
 }
